changed return types to Result where it possible

+++ FiestaMarketBackend.Application/Product/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -1,10 +1,11 @@
-﻿using FiestaMarketBackend.Application.Services;
+﻿using CSharpFunctionalExtensions;
+using FiestaMarketBackend.Application.Services;
 using FiestaMarketBackend.Infrastructure.Repositories;
 using MediatR;
 
 namespace FiestaMarketBackend.Application.Product.Commands
 {
-    public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand>
+    public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand, Result>
     {
         private readonly ProductsRepository _productsRepository;
         private readonly FileService _fileService;
@@ -15,13 +16,24 @@ namespace FiestaMarketBackend.Application.Product.Commands
             _fileService = fileService;
         }
 
-        public async Task Handle(DeleteProductCommand request, CancellationToken cancellationToken)
+        public async Task<Result> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
-            var images = await _productsRepository.GetImagesAsync(request.Id);
-            await _productsRepository.DeleteAsync(request.Id);
-            _fileService.DeleteImages(images);
+            var imagesResult = await _productsRepository.GetImagesAsync(request.Id);
 
-            return;
+            if (imagesResult.IsFailure)
+                return Result.Failure(imagesResult.Error);
+
+            var repoResult = await _productsRepository.DeleteAsync(request.Id);
+
+            if (repoResult.IsFailure) 
+                return Result.Failure(repoResult.Error);
+
+            var fileResult = _fileService.DeleteImages(imagesResult.Value);
+
+            if (fileResult.IsFailure)
+                return Result.Failure(fileResult.Error);
+
+            return Result.Success();
         }
     }
 }
